Fix oxygen consumption code (idPlayer::airTics) for > 60Hz
I turned idPlayer::airTics into a float (so they are now virtual
 60Hz tics instead of actual number of tics) and adjusted the code
that increases or decreases the airTics (idPlayer::UpdateAir())
to scale the values added/subtracted accordingly.

+++ neo/game/Player.h
@@ -564,7 +564,8 @@ private:
 	int						numProjectileHits;		// number of hits on mobs
 
 	bool					airless;
-	int						airTics;				// set to pm_airTics at start, drops in vacuum
+	// DG: Note: airTics are tics at 60Hz, so no real tics (unless com_gameHz happens to be 60)
+	float					airTics;				// set to pm_airTics at start, drops in vacuum
 	int						lastAirDamage;
 
 	bool					gibDeath;
